✅ a bit smoother

+++ Exercises/501-1000/726-750/747SmallestIndexEqValue/smIdxEqValue.rs
@@ -57,8 +57,8 @@ Acceptance Rate
 impl Solution {
   pub fn smallest_equal(nums: Vec<i32>) -> i32 {
     for i in 0..nums.len() {
-      let md = i % 10;
-      if md == nums[i] { return i }
+      let md = i as i32 % 10;
+      if md == nums[i] { return i as i32 }
     }
     -1
   }
